+++ core/src/main/java/cn/leancloud/AVObject.java
@@ -34,6 +34,8 @@ public class AVObject {
 
   public static final String KEY_CLASSNAME = "className";
 
+  public static final String KEY_IGNORE_HOOKS = "__ignore_hooks";
+
   private static final String INTERNAL_PATTERN = "^[\\da-z][\\d-a-z]*$";
   private static final Set<String> RESERVED_ATTRS = new HashSet<String>(
           Arrays.asList(KEY_CREATED_AT, KEY_UPDATED_AT, KEY_OBJECT_ID, KEY_ACL));
@@ -52,6 +54,12 @@ public class AVObject {
 
   private volatile boolean fetchWhenSave = false;
   protected volatile boolean totallyOverwrite = false;
+
+  public enum Hook {
+    beforeSave, afterSave, beforeUpdate, afterUpdate, beforeDelete, afterDelete,
+  }
+
+  private transient Set<Hook> ignoreHooks = new TreeSet<Hook>();
 
   /**
    * Default constructor.
@@ -705,6 +713,11 @@ public class AVObject {
       tmp.remove(KEY_CREATED_AT);
       tmp.remove(KEY_UPDATED_AT);
       tmp.remove(KEY_OBJECT_ID);
+
+      if (ignoreHooks.size() > 0) {
+        tmp.put(KEY_IGNORE_HOOKS, ignoreHooks);
+      }
+ 
       return JSONObject.Builder.create(tmp);
     }
 
@@ -723,6 +736,10 @@ public class AVObject {
         ObjectFieldOperation op = OperationBuilder.gBuilder.create(OperationBuilder.OperationType.Set, KEY_ACL, acl);
         params.putAll(op.encode());
       }
+    }
+
+    if (ignoreHooks.size() > 0) {
+      params.put(KEY_IGNORE_HOOKS, ignoreHooks);
     }
 
     if (!needBatchMode()) {
@@ -1145,10 +1162,14 @@ public class AVObject {
    * @return observable instance.
    */
   public Observable<AVNull> deleteInBackground() {
-    if (totallyOverwrite) {
-      return PaasClient.getStorageClient().deleteWholeObject(this.endpointClassName, getObjectId());
+    Map<String, Object> ignoreParam = new HashMap<>();
+    if (ignoreHooks.size() > 0) {
+      ignoreParam.put(KEY_IGNORE_HOOKS, ignoreHooks);
     }
-    return PaasClient.getStorageClient().deleteObject(this.className, getObjectId());
+    if (totallyOverwrite) {
+      return PaasClient.getStorageClient().deleteWholeObject(this.endpointClassName, getObjectId(), ignoreParam);
+    }
+    return PaasClient.getStorageClient().deleteObject(this.className, getObjectId(), ignoreParam);
   }
 
   /**
@@ -1177,6 +1198,7 @@ public class AVObject {
       return Observable.just(AVNull.getINSTANCE());
     }
     String className = null;
+    Map<String, Object> ignoreParams = new HashMap<>();
     StringBuilder sb = new StringBuilder();
     for (AVObject o : objects) {
       if (StringUtil.isEmpty(o.getObjectId()) || StringUtil.isEmpty(o.getClassName())) {
@@ -1191,7 +1213,7 @@ public class AVObject {
         return Observable.error(new IllegalArgumentException("The objects class name must be the same."));
       }
     }
-    return PaasClient.getStorageClient().deleteObject(className, sb.toString());
+    return PaasClient.getStorageClient().deleteObject(className, sb.toString(), ignoreParams);
   }
 
   /**
@@ -1524,6 +1546,28 @@ public class AVObject {
     }
   }
 
+  /**
+   * disable beforeXXX Hooks
+   */
+  public void disableBeforeHook() {
+    Collections.addAll(ignoreHooks, Hook.beforeSave, Hook.beforeUpdate, Hook.beforeDelete);
+  }
+
+  /**
+   * disable afterXXX Hooks
+   */
+  public void disableAfterHook() {
+    Collections.addAll(ignoreHooks, Hook.afterSave, Hook.afterUpdate, Hook.afterDelete);
+  }
+
+  /**
+   * ignore specified Hook
+   * @param hook target Hook.
+   */
+  public void ignoreHook(Hook hook) {
+    ignoreHooks.add(hook);
+  }
+
   protected static <T extends AVObject> T cast(AVObject object, Class<T> clazz) throws Exception {
     if (clazz.getClass().isAssignableFrom(object.getClass())) {
       return (T) object;
@@ -1558,7 +1602,6 @@ public class AVObject {
 
   @Override
   public int hashCode() {
-
     return Objects.hash(getClassName(), getServerData(), operations, acl, isFetchWhenSave());
   }
 }