+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 Following is change logs for recently release versions, you can refer to [releases page](https://github.com/leancloud/java-unified-sdk/releases) for more details.
 
+## 6.5.8 release
+
+#### Break changes
+- None
+
+#### New features
+- support disable hook for leanengine.
+
+#### Optimization and fixed bugs
+- fixed: AddUniqueOperation#apply error for AbstractList.
+
 ## 6.5.7 release
 
 #### Break changes